kernel: armksyms.c: Remove strlen and strncpy symbols from user
So that user can't misuse them while debugging

+++ kernel/rcutree.c
@@ -1529,7 +1529,7 @@ static void rcu_do_batch(struct rcu_state *rsp, struct rcu_data *rdp)
 {
 	unsigned long flags;
 	struct rcu_head *next, *list, **tail;
-	int bl, count, count_lazy;
+	long bl, count, count_lazy;
 	int i;
 
 	/* If no callbacks are ready, just return.*/
